+++ frontend/src/views/files/plyrViewer.vue
@@ -138,7 +138,7 @@
           @play="handlePlay"
           playsinline
         >
-          <track kind="captions" v-for="(sub, index) in subtitlesList" :key="index" :src="sub.src"
+          <track kind="captions" v-for="(sub, index) in subtitlesList" :key="sub.language || index" :src="sub.src"
             :label="subtitleTrackLabel(sub)" :srclang="sub.language" />
         </video>
       </div>